feature/OMP-US-10: в сервисный слой добавлено решение вопроса уникальности для UserDto

+++ src/main/java/edu/online/messenger/service/impl/UserServiceImpl.java
@@ -77,7 +77,7 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public AddressDto addAddressToUser(AddressCreateDto addressCreateDto) {
-        Long userId = (long) addressCreateDto.getUserId();
+        Long userId = addressCreateDto.getUserId();
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(userId));
         Address address = addressMapper.toEntity(addressCreateDto);
